% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPCmodel.R
\name{get_required_cpgs}
\alias{get_required_cpgs}
\title{Get Required CpG Sites}
\usage{
get_required_cpgs()
}
\value{
A character vector of CpG site IDs.
}
\description{
Returns the list of CpG sites required for the calculation of epigenetic gestational age.
}
